Renamed toolbar to MenuBar
- fix styles
